+++ EmployeeLeaveManagementWebAPI/DAL/EmployeeLeaveTransaction.cs
@@ -26,7 +26,7 @@ namespace LMS_WebAPI_DAL
         public System.DateTime ToDate { get; set; }
         public System.DateTime CreatedDate { get; set; }
         public int RefStatus { get; set; }
-        public int NumberOfWorkingDays { get; set; }
+        public double NumberOfWorkingDays { get; set; }
         public int RefLeaveType { get; set; }
         public string EmployeeComment { get; set; }
         public Nullable<System.DateTime> ModifiedDate { get; set; }
